requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
yt-dlp==2023.11.16
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
supabase==2.3.4
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
